requirements.txt
faiss-cpu
sentence-transformers
transformers
torch
datasets
numpy
deep-translator
